% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_cpr.R
\name{calc_so_cpr_abundance_index}
\alias{calc_so_cpr_abundance_index}
\title{Calculate the segment-corrected total plankton abundance index}
\usage{
calc_so_cpr_abundance_index(cpr_data)
}
\arguments{
\item{cpr_data}{Data frame. Raw SO-CPR data, e.g.
\code{download_so_cpr_data()$data}.}
}
\value{
A list with elements:
\describe{
\item{data}{\code{cpr_data} with an added \code{Total_Plankton_Segment_corr} column}
\item{effort}{A data frame of sample counts per year}
}
}
\description{
Calculate the segment-corrected total plankton abundance index
}
\examples{
\dontrun{
cpr <- download_so_cpr_data()
abund <- calc_so_cpr_abundance_index(cpr$data)
}

}
\seealso{
\code{\link[=plot_so_cpr_abundance_index]{plot_so_cpr_abundance_index()}}
}
