% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_cpr.R
\name{plot_so_cpr_salp_euph_index}
\alias{plot_so_cpr_salp_euph_index}
\title{Plot the Salp/Euphausiid abundance balance over time}
\usage{
plot_so_cpr_salp_euph_index(salp_euph_list, ylim = c(-50, 50))
}
\arguments{
\item{salp_euph_list}{List. Output of \code{\link[=calc_so_cpr_salp_euph_index]{calc_so_cpr_salp_euph_index()}}.}

\item{ylim}{Numeric vector of length 2. y-axis limits for the
segment-level balance. Defaults to \code{c(-50, 50)}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot the Salp/Euphausiid abundance balance over time
}
