% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_cpr.R
\name{so_cpr_segment_id}
\alias{so_cpr_segment_id}
\title{Build a unique segment identifier for SO-CPR samples}
\usage{
so_cpr_segment_id(cpr_data)
}
\arguments{
\item{cpr_data}{Data frame. Raw SO-CPR data (must contain
\code{Ship_Code}, \code{Tow_Number}, \code{Date}, \code{Time}).}
}
\value{
\code{cpr_data} with an added character column \code{Segment}.
}
\description{
Build a unique segment identifier for SO-CPR samples
}
