% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soae-package.R
\docType{package}
\name{soae-package}
\alias{soae}
\alias{soae-package}
\title{soae: State of the Antarctic Environment (soae) package}
\description{
This package can be used to render the figures for the State of the Antarctic Environment (SOAE) report.
}
\author{
\strong{Maintainer}: Pablo Deschepper \email{pdeschepper@naturalsciences.be}

Authors:
\itemize{
  \item Ben Raymond
  \item Mike Meredyth-Young
}

}
\keyword{internal}
